fix: moves buttons to extra classes
