--- src/main/java/com/github/zfl4wless/ui/pages/listeners/StartButtonActionListener.java
@@ -1,11 +1,0 @@
-package com.github.zfl4wless.ui.pages.listeners;
-
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-public class StartButtonActionListener implements ActionListener {
-    @Override
-    public void actionPerformed(ActionEvent e) {
-        //
-    }
-}
